добавлен dll WebServerCore

+++ Chat_Server/Old/ClientClass.cs
@@ -8,10 +8,8 @@ using System.Threading;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
 
-namespace Chat_server
-{
-    class ClientClass
-    {
+namespace Chat_server {
+    class ClientClass {
         public static int Count { get; private set; }
 
         private TcpClient tcpClient;
@@ -19,15 +17,14 @@ namespace Chat_server
         private bool useWebsocket;
         public byte[] UniqueId { get; private set; }
         public string Login { get; private set; }
-        public long Id { get; private set; }
+        public uint Id { get; private set; }
         public ClientStatus Status { get; private set; }
         private object locker;
 
         //DEBUG
         public string Address { get; private set; }
-        
-        public ClientClass(TcpClient tcpClient, TCP_Server tcpServer)
-        {
+
+        public ClientClass(TcpClient tcpClient, TCP_Server tcpServer) {
             this.tcpClient = tcpClient;
             this.tcpServer = tcpServer;
             UniqueId = Guid.NewGuid().ToByteArray();
@@ -39,40 +36,33 @@ namespace Chat_server
             Address = tcpClient.Client.RemoteEndPoint.ToString();
         }
 
-        public void Start()
-        {
+        public void Start() {
             Status = ClientStatus.Start;
             Task.Factory.StartNew(Process, TaskCreationOptions.LongRunning);
         }
 
-        public void Close()
-        {
+        public void Close() {
             tcpClient.Close();
         }
 
-        private void Process()
-        {
-            if (tcpClient.Connected)
-            {
+        private void Process() {
+            if (tcpClient.Connected) {
                 Console.WriteLine(DateTime.Now + " [DEBUG][TCP] " + tcpClient.Client.RemoteEndPoint + " подключился");
                 //АВТОРИЗАЦИЯ
                 Status = ClientStatus.Authorization;
-                try
-                {
+                try {
                     if (Authorization()) Status = ClientStatus.Logged;
                 }
-                catch (Exception)
-                {
+                catch (Exception) {
                     tcpServer.ClientClosed(this);
                     return;
                 }
-                if (Status == ClientStatus.Logged)
-                {
+                if (Status == ClientStatus.Logged) {
                     //if (Program.DEBUG) Console.WriteLine(DateTime.Now + " [DEBUG][TCP] " + tcpClient.Client.RemoteEndPoint + " авторизовался как " + Login);
 
                     Status = ClientStatus.LoadData;
                     {
-                        Send(JsonConvert.SerializeObject(new CW("dat", JsonConvert.SerializeObject(new CW.Data{
+                        Send(JsonConvert.SerializeObject(new CW("dat", JsonConvert.SerializeObject(new CW.Data {
                             login = Login,
                             online = tcpServer.Online
                         }))));
@@ -81,44 +71,36 @@ namespace Chat_server
                             for (int i = 0; i < history.Length; i++)
                                 Send(JsonConvert.SerializeObject(new CW("mes", JsonConvert.SerializeObject(history[i]))));
                     }
-                    
+
                     Status = ClientStatus.Ready;
 
                     Count++;
                     Task.Factory.StartNew(() => tcpServer.ClientLogged(this));
 
-                    try
-                    {
-                        while (true)
-                        {
+                    try {
+                        while (true) {
                             string inputData = Read();
-                            if (inputData.Length == 2 && inputData[0] == 3 && inputData[1] == 65533)
-                            {
+                            if (inputData.Length == 2 && inputData[0] == 3 && inputData[1] == 65533) {
                                 Count--;
                                 Status = ClientStatus.Stopped;
                                 tcpServer.ClientClosed(this);
                                 return;
                             }
-                            else if (inputData.Length > 1500)
-                            {
+                            else if (inputData.Length > 1500) {
                                 continue;
                                 /*string json = JsonConvert.SerializeObject(new CW("err", "LEN"));*/
                             }
-                            else if (inputData.Length > 20)
-                            {
+                            else if (inputData.Length > 20) {
                                 CW wrap = JsonConvert.DeserializeObject<CW>(inputData);
-                                if (wrap.type == "mes")
-                                {
+                                if (wrap.type == "mes") {
                                     CW.MessageFromClient json = JsonConvert.DeserializeObject<CW.MessageFromClient>(wrap.body);
                                     if (json.message.Length > 0)
                                         Task.Factory.StartNew(() => tcpServer.Message(this, json.to, json.message));
                                     //if (Program.DEBUG)
                                     //Console.WriteLine(DateTime.Now + " [DEBUG][TCP] " + Login + ": " + json.message.Replace());
                                 }
-                                else if (wrap.type == "getm")
-                                {
-                                    Task.Factory.StartNew(() =>
-                                    {
+                                else if (wrap.type == "getm") {
+                                    Task.Factory.StartNew(() => {
                                         CW.MessageToClient[] history = MySql_Client.LoadMessagePrivate(Id, wrap.body);
                                         if (history != null)
                                             for (int i = 0; i < history.Length; i++)
@@ -128,8 +110,7 @@ namespace Chat_server
                             }
                         }
                     }
-                    catch (System.IO.IOException)
-                    {
+                    catch (System.IO.IOException) {
                         Count--;
                         Status = ClientStatus.Stopped;
                         tcpServer.ClientClosed(this);
@@ -141,12 +122,10 @@ namespace Chat_server
             else tcpServer.ClientClosed(this);
         }
 
-        private bool Authorization()
-        {
+        private bool Authorization() {
             string inputData = ReadHttp();
             if (inputData.Length > 15 && inputData.Substring(0, 14) == "GET /websocket"
-                && Regex.IsMatch(inputData, "Connection: Upgrade") && Regex.IsMatch(inputData, "Upgrade: websocket"))
-            {
+                && Regex.IsMatch(inputData, "Connection: Upgrade") && Regex.IsMatch(inputData, "Upgrade: websocket")) {
                 SendHttp("HTTP/1.1 101 Switching Protocols" + Environment.NewLine + "Connection: Upgrade" + Environment.NewLine
                     + "Upgrade: websocket" + Environment.NewLine + "Sec-WebSocket-Accept: " + Convert.ToBase64String(
                         System.Security.Cryptography.SHA1.Create().ComputeHash(Encoding.UTF8.GetBytes(
@@ -154,34 +133,26 @@ namespace Chat_server
                                 + Environment.NewLine + Environment.NewLine);
                 useWebsocket = true;
             }
-            else if (inputData == "CONNECT STANDART TCP")
-            {
+            else if (inputData == "CONNECT STANDART TCP") {
                 useWebsocket = false;
                 Send("CONNECTION OPEN");
             }
-            else
-            {
+            else {
                 Console.WriteLine(DateTime.Now + " [DEBUG][TCP] " + tcpClient.Client.RemoteEndPoint + " прислал ЭТО " + inputData);
                 SendHttp("ERROR-FORMAT");
                 return false;
             }
-            while (true)
-            {
+            while (true) {
                 inputData = Read();
                 CW wrap = JsonConvert.DeserializeObject<CW>(inputData);
-                if (wrap.type != null && wrap.body != null)
-                {
-                    if (wrap.type == "log")
-                    {
+                if (wrap.type != null && wrap.body != null) {
+                    if (wrap.type == "log") {
                         CW.Login loginJSON = JsonConvert.DeserializeObject<CW.Login>(wrap.body);
-                        if (loginJSON.login != null && loginJSON.pass != null && loginJSON.login.Length >= 4 && loginJSON.pass.Length >= 4)
-                        {
-                            if (loginJSON.withpass)
-                            {
+                        if (loginJSON.login != null && loginJSON.pass != null && loginJSON.login.Length >= 4 && loginJSON.pass.Length >= 4) {
+                            if (loginJSON.withpass) {
                                 string[] res = MySql_Client.CheckUser(loginJSON.login, loginJSON.pass);
-                                if (res != null)
-                                {
-                                    Id = Convert.ToInt64(res[0]);
+                                if (res != null) {
+                                    Id = Convert.ToUInt32(res[0]);
                                     Login = res[1];
                                     Send("LOGINED-SUCCSESS-ENTER-CHAT");
                                     return true;
@@ -190,23 +161,19 @@ namespace Chat_server
                             }
                             else Send("NOT-WORK-USE-WITHPASS");
                         }
-                        else
-                        {
+                        else {
                             Send("ERROR-DATA-CONNECTION-CLOSING");
                             return false;
                         }
                     }
-                    else if (wrap.type == "reg")
-                    {
+                    else if (wrap.type == "reg") {
                         CW.Registration regJSON = JsonConvert.DeserializeObject<CW.Registration>(wrap.body);
                         if (regJSON.login != null && regJSON.pass != null && regJSON.email != null)
                             if (Regular.CheckLogin(regJSON.login))
                                 if (Regular.CheckPass(regJSON.pass))
-                                    if (Regular.CheckEmail(regJSON.email))
-                                    {
+                                    if (Regular.CheckEmail(regJSON.email)) {
                                         Id = MySql_Client.RegNewUser(regJSON.login, regJSON.pass, regJSON.email);
-                                        if (Id > 0)
-                                        {
+                                        if (Id > 0) {
                                             Login = regJSON.login;
                                             Send("REGISTR-SUCCSESS-ENTER-CHAT");
                                             return true;
@@ -216,31 +183,27 @@ namespace Chat_server
                                     else Send("ERR-REGEMAIL");
                                 else Send("ERR-REGPASS");
                             else Send("ERR-REGLOGIN");
-                        else
-                        {
+                        else {
                             Send("ERROR-DATA-CONNECTION-CLOSING");
                             return false;
                         }
                     }
-                    else
-                    {
+                    else {
                         Console.WriteLine(inputData);
                         Send("ERROR-DATA-CONNECTION-CLOSING");
                         return false;
                     }
                 }
-                else
-                {
+                else {
                     Send("ERROR-DATA-CONNECTION-CLOSING");
                     return false;
                 }
             }
         }
 
-        public void SendMessage(long id, string sender, DateTime date, string message, string recipient = "public")
-        {
+        public void SendMessage(long id, string sender, DateTime date, string message, string recipient = "public") {
             Send(JsonConvert.SerializeObject(new CW("mes",
-                JsonConvert.SerializeObject(new CW.MessageToClient(){
+                JsonConvert.SerializeObject(new CW.MessageToClient() {
                     id = id,
                     date = date,
                     from = sender,
@@ -249,22 +212,18 @@ namespace Chat_server
                 }))));
         }
 
-        public void Send_Online(string login)
-        {
+        public void Send_Online(string login) {
             Send(JsonConvert.SerializeObject(
                 new CW("onl", login)));
         }
 
-        public void Send_Offline(string login)
-        {
+        public void Send_Offline(string login) {
             Send(JsonConvert.SerializeObject(
                 new CW("off", login)));
         }
 
-        private string Read()
-        {
-            if (useWebsocket)
-            {
+        private string Read() {
+            if (useWebsocket) {
                 byte[] buf = new byte[10000];
                 int readed = tcpClient.GetStream().Read(buf, 0, buf.Length);
                 byte[] inputData = new byte[readed];
@@ -287,14 +246,11 @@ namespace Chat_server
 
                 return Encoding.UTF8.GetString(decoded, 0, decoded.Length);
             }
-            else
-            {
+            else {
                 byte[] buf = new byte[10000];
-                while (true)
-                {
+                while (true) {
                     int readed = tcpClient.GetStream().Read(buf, 0, buf.Length);
-                    if (readed > 0)
-                    {
+                    if (readed > 0) {
                         byte[] inputData = new byte[readed];
                         for (long i = 0; i < readed; i++)
                             inputData[i] = buf[i];
@@ -304,30 +260,25 @@ namespace Chat_server
                 }
             }
         }
-        private void Send(string message)
-        {
-            if (useWebsocket)
-            {
+        private void Send(string message) {
+            if (useWebsocket) {
                 byte[] outputData;
                 byte[] bytesRaw = Encoding.UTF8.GetBytes(message);
                 byte[] frame = new byte[10];
                 int indexStartRawData = -1;
                 int length = bytesRaw.Length;
                 frame[0] = (byte)129;
-                if (length <= 125)
-                {
+                if (length <= 125) {
                     frame[1] = (byte)length;
                     indexStartRawData = 2;
                 }
-                else if (length >= 126 && length <= 65535)
-                {
+                else if (length >= 126 && length <= 65535) {
                     frame[1] = (byte)126;
                     frame[2] = (byte)((length >> 8) & 255);
                     frame[3] = (byte)(length & 255);
                     indexStartRawData = 4;
                 }
-                else
-                {
+                else {
                     frame[1] = (byte)127;
                     frame[2] = (byte)((length >> 56) & 255);
                     frame[3] = (byte)((length >> 48) & 255);
@@ -350,15 +301,13 @@ namespace Chat_server
                 lock (locker)
                     tcpClient.GetStream().Write(outputData, 0, outputData.Length);
             }
-            else
-            {
+            else {
                 byte[] outputData = Encoding.UTF8.GetBytes(message + "\r\n");
                 tcpClient.GetStream().Write(outputData, 0, outputData.Length);
             }
         }
 
-        private string ReadHttp()
-        {
+        private string ReadHttp() {
             byte[] buf = new byte[10000000];
             int readed = tcpClient.GetStream().Read(buf, 0, buf.Length);
             byte[] inputData = new byte[readed];
@@ -366,14 +315,12 @@ namespace Chat_server
                 inputData[i] = buf[i];
             return Encoding.UTF8.GetString(inputData);
         }
-        private void SendHttp(string outputData)
-        {
+        private void SendHttp(string outputData) {
             byte[] outputB = Encoding.UTF8.GetBytes(outputData);
             tcpClient.GetStream().Write(outputB, 0, outputB.Length);
         }
 
-        public enum ClientStatus
-        {
+        public enum ClientStatus {
             Created,
             Start,
             Authorization,
@@ -383,47 +330,40 @@ namespace Chat_server
             Stopped
         }
 
-        public class CW
-        {
+        public class CW {
             public string type { get; set; }
             public string body { get; set; }
 
-            public CW(string mestype, string body)
-            {
+            public CW(string mestype, string body) {
                 this.type = mestype;
                 this.body = body;
             }
 
-            public class Login
-            {
+            public class Login {
                 public string login { get; set; }
                 public bool withpass { get; set; }
                 public string pass { get; set; }
                 public string key { get; set; }
             }
 
-            public class Registration
-            {
+            public class Registration {
                 public string login { get; set; }
                 public string pass { get; set; }
                 public string email { get; set; }
             }
 
-            public class MessageFromClient
-            {
+            public class MessageFromClient {
                 public string to { get; set; }
                 public string message { get; set; }
             }
 
-            public class MessageToClient : IComparable
-            {
+            public class MessageToClient : IComparable {
                 public long id { get; set; }
                 public string from { get; set; }
                 public string to { get; set; }
                 public DateTime date { get; set; }
                 public string message { get; set; }
-                public int CompareTo(object o)
-                {
+                public int CompareTo(object o) {
                     MessageToClient m = o as MessageToClient;
                     if (m != null)
                         return id.CompareTo(m.id);
@@ -432,8 +372,7 @@ namespace Chat_server
                 }
             }
 
-            public class Data
-            {
+            public class Data {
                 public string login { get; set; }
                 public string[] online { get; set; }
             }
